rename swift and wasm crates
