+++ .github/workflows/ci.yml
@@ -141,7 +141,7 @@ jobs:
         run: curl -LsSf https://get.nexte.st/latest/linux | tar zxf - -C ${CARGO_HOME:-~/.cargo}/bin
 
       - name: Run unit tests with coverage
-        run: cargo llvm-cov nextest --profile ci --locked --lib --workspace --no-fail-fast --all-features --cobertura --output-path cobertura.xml --exclude etopay-sdk-jni --exclude cawaena-sdk-swift --exclude cawaena-sdk-wasm --ignore-filename-regex "sdk/bindings/|api_types/|.*/error\.rs"
+        run: cargo llvm-cov nextest --profile ci --locked --lib --workspace --no-fail-fast --all-features --cobertura --output-path cobertura.xml --exclude etopay-sdk-jni --exclude etopay-sdk-swift --exclude etopay-sdk-wasm --ignore-filename-regex "sdk/bindings/|api_types/|.*/error\.rs"
 
       - name: Publish Coverage
         uses: 5monkeys/cobertura-action@master
@@ -190,7 +190,7 @@ jobs:
       - name: Upload Artifact
         uses: actions/upload-artifact@v4
         with:
-          name: cawaena-sdk-wasm
+          name: etopay-sdk-wasm
           path: sdk/bindings/wasm/out/**/*.tgz
           if-no-files-found: error
           retention-days: 7
