+++ .github/workflows/deploy-wasm.yml
@@ -54,7 +54,7 @@ jobs:
       - name: Upload Artifact
         uses: actions/upload-artifact@v4
         with:
-          name: cawaena-sdk-wasm
+          name: etopay-sdk-wasm
           path: sdk/bindings/wasm/out/**/*.tgz
           if-no-files-found: error
           retention-days: 7
