+++ Makefile
@@ -2,7 +2,7 @@ build:
 	cargo build --release --locked
 	
 code-coverage:
-	cargo llvm-cov nextest --release --locked --all-features --lib --workspace --html --exclude etopay-sdk-jni --exclude cawaena-sdk-swift --exclude cawaena-sdk-wasm --ignore-filename-regex "sdk/bindings/|api_types/|.*/error\.rs" --no-fail-fast
+	cargo llvm-cov nextest --release --locked --all-features --lib --workspace --html --exclude etopay-sdk-jni --exclude etopay-sdk-swift --exclude etopay-sdk-wasm --ignore-filename-regex "sdk/bindings/|api_types/|.*/error\.rs" --no-fail-fast
 
 unit-tests:
 	cargo nextest run --profile ci --release --locked --lib --no-fail-fast
