+++ sdk/bindings/swift/Cargo.toml
@@ -7,7 +7,7 @@ homepage.workspace = true
 repository.workspace = true
 publish = false
 
-name = "cawaena-sdk-swift"
+name = "etopay-sdk-swift"
 description = "Swift bindings for Cawaena SDK"
 build = "build.rs"
