+++ sdk/bindings/swift/README.md
@@ -42,8 +42,8 @@ Our xcframework structure is as following:
   - ios-arm64/
     - Headers/
       - generated/
-        - cawaena-sdk-swift/
-          - cawaena-sdk-swift.h
+        - etopay-sdk-swift/
+          - etopay-sdk-swift.h
         - SwiftBridgeCore.h
       - bridging-header.h
       - module.modulemap
@@ -78,11 +78,11 @@ Previously we were bundling our Swift bindings into a `.zip` file and exporting 
   - Sources/
     - CawaenaSdk/
       - SwiftBridgeCore.swift
-      - cawaena-sdk-swift.swift
+      - etopay-sdk-swift.swift
       - utils.swift
   - Package.swift
 
-The `CawaenaSdkBin.xcframework` creation and its internal structure is explained in the above `Generating iOS xcframeworks` section. The `SwiftBridgeCore.swift` and `cawaena-sdk-swift.swift` files are auto-generated by the `swift-bridge` crate.
+The `CawaenaSdkBin.xcframework` creation and its internal structure is explained in the above `Generating iOS xcframeworks` section. The `SwiftBridgeCore.swift` and `etopay-sdk-swift.swift` files are auto-generated by the `swift-bridge` crate.
 
 Finally, the `Package.swift` is the actual file that generates the Swift package called `CawaenaSdk`. In xcode we can add the CawaenaSdk package and initialize the constructor and call its functions (our Rust bindings):
