+++ sdk/bindings/wasm/Cargo.toml
@@ -7,7 +7,7 @@ homepage.workspace = true
 repository.workspace = true
 publish = false
 
-name = "cawaena-sdk-wasm"
+name = "etopay-sdk-wasm"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [lints]
