+++ sdk/bindings/wasm/examples/02-onboard_user_via_postident.ts
@@ -1,5 +1,5 @@
 //  Need to do manual verification on postident: https://postident-itu.deutschepost.de/testapp
-import * as wasm from "../pkg/cawaena_sdk_wasm";
+import * as wasm from "../pkg/etopay_sdk_wasm";
 import { initSdk } from './utils';
 
 async function main() {
