+++ sdk/bindings/wasm/examples/03-create_new_wallet.ts
@@ -1,4 +1,4 @@
-import * as wasm from "../pkg/cawaena_sdk_wasm";
+import * as wasm from "../pkg/etopay_sdk_wasm";
 import { initSdk } from './utils';
 
 async function main() {
