+++ sdk/bindings/wasm/examples/07-get_balance.ts
@@ -1,4 +1,4 @@
-import * as wasm from "../pkg/cawaena_sdk_wasm";
+import * as wasm from "../pkg/etopay_sdk_wasm";
 import { initSdk } from './utils';
 
 async function main() {
