+++ sdk/bindings/wasm/examples/16-get_tx_list.ts
@@ -1,4 +1,4 @@
-import * as wasm from "../pkg/cawaena_sdk_wasm";
+import * as wasm from "../pkg/etopay_sdk_wasm";
 import { initSdk } from './utils';
