+++ sdk/bindings/wasm/examples/webapp/examples/02-onboard_user_via_postident.js
@@ -1,5 +1,5 @@
 //  Need to do manual verification on postident: https://postident-itu.deutschepost.de/testapp
-import * as wasm from "cawaena-sdk-wasm";
+import * as wasm from "etopay-sdk-wasm";
 import { initSdk } from './util';
 
 async function main() {
