+++ sdk/bindings/wasm/examples/webapp/examples/20-send_compliment.js
@@ -1,4 +1,4 @@
-import * as wasm from "cawaena-sdk-wasm";
+import * as wasm from "etopay-sdk-wasm";
 import { sleep, initSdk } from "./util";
 
 async function main() {
